# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Provena",
    "summary": "Provenance-first memory service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/memories": {
      "post": {
        "summary": "Create Memory",
        "operationId": "create_memory_v1_memories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryWriteResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/memories/search": {
      "post": {
        "summary": "Search Memories",
        "operationId": "search_memories_v1_memories_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ACLEntry": {
        "properties": {
          "principal_id": {
            "type": "string",
            "title": "Principal Id"
          },
          "principal_type": {
            "type": "string",
            "title": "Principal Type"
          },
          "permissions": {
            "items": {
              "$ref": "#/components/schemas/ACLPermission"
            },
            "type": "array",
            "title": "Permissions"
          }
        },
        "type": "object",
        "required": [
          "principal_id",
          "principal_type"
        ],
        "title": "ACLEntry"
      },
      "ACLPermission": {
        "type": "string",
        "enum": [
          "read",
          "write",
          "delete",
          "share",
          "admin"
        ],
        "title": "ACLPermission"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemoryCreate": {
        "properties": {
          "memory_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Id"
          },
          "kind": {
            "$ref": "#/components/schemas/MemoryKind"
          },
          "scope": {
            "$ref": "#/components/schemas/ScopeEnvelope"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "entity_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Entity Keys"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "importance": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Importance",
            "default": 0.5
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 0.7
          },
          "strength": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Strength",
            "default": 0.7
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          },
          "source_references": {
            "items": {
              "$ref": "#/components/schemas/SourceReference"
            },
            "type": "array",
            "title": "Source References"
          },
          "supersedes_memory_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Memory Id"
          },
          "trigger_phrases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Trigger Phrases"
          },
          "acl": {
            "items": {
              "$ref": "#/components/schemas/ACLEntry"
            },
            "type": "array",
            "title": "Acl"
          },
          "embedding_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Model"
          },
          "embedding": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Embedding"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "scope",
          "content"
        ],
        "title": "MemoryCreate"
      },
      "MemoryKind": {
        "type": "string",
        "enum": [
          "fact",
          "episode",
          "artifact",
          "decision",
          "relation",
          "preference",
          "instruction",
          "metric",
          "stakeholder",
          "source"
        ],
        "title": "MemoryKind"
      },
      "MemoryRecord": {
        "properties": {
          "memory_id": {
            "type": "string",
            "title": "Memory Id"
          },
          "fingerprint": {
            "type": "string",
            "title": "Fingerprint"
          },
          "kind": {
            "$ref": "#/components/schemas/MemoryKind"
          },
          "status": {
            "$ref": "#/components/schemas/MemoryStatus"
          },
          "scope": {
            "$ref": "#/components/schemas/ScopeEnvelope"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "entity_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Entity Keys"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "importance": {
            "type": "number",
            "title": "Importance"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "strength": {
            "type": "number",
            "title": "Strength"
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "last_verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified At"
          },
          "source_references": {
            "items": {
              "$ref": "#/components/schemas/SourceReference"
            },
            "type": "array",
            "title": "Source References"
          },
          "acl": {
            "items": {
              "$ref": "#/components/schemas/ACLEntry"
            },
            "type": "array",
            "title": "Acl"
          },
          "trigger_phrases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Trigger Phrases"
          },
          "embedding_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Model"
          },
          "embedding": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Embedding"
          },
          "held": {
            "type": "boolean",
            "title": "Held",
            "default": false
          },
          "hold_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hold Reason"
          },
          "hold_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hold Until"
          }
        },
        "type": "object",
        "required": [
          "memory_id",
          "fingerprint",
          "kind",
          "status",
          "scope",
          "content",
          "importance",
          "confidence",
          "strength",
          "created_at",
          "updated_at"
        ],
        "title": "MemoryRecord"
      },
      "MemoryStatus": {
        "type": "string",
        "enum": [
          "active",
          "superseded",
          "deleted",
          "held"
        ],
        "title": "MemoryStatus"
      },
      "MemoryWriteResult": {
        "properties": {
          "created": {
            "type": "boolean",
            "title": "Created"
          },
          "memory": {
            "$ref": "#/components/schemas/MemoryRecord"
          }
        },
        "type": "object",
        "required": [
          "created",
          "memory"
        ],
        "title": "MemoryWriteResult"
      },
...
      "RelatedMemory": {
        "properties": {
          "relation": {
            "$ref": "#/components/schemas/RelationKind"
          },
          "memory": {
            "$ref": "#/components/schemas/MemoryRecord"
          }
        },
        "type": "object",
        "required": [
          "relation",
          "memory"
        ],
        "title": "RelatedMemory"
      },
      "RelationKind": {
        "type": "string",
        "enum": [
          "related_to",
          "supports",
          "derived_from",
          "supersedes",
          "conflicts_with"
        ],
        "title": "RelationKind"
      },
...
      "ScopeEnvelope": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "tenant_id"
        ],
        "title": "ScopeEnvelope"
      },
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "scope": {
            "$ref": "#/components/schemas/ScopeEnvelope"
          },
          "kinds": {
            "items": {
              "$ref": "#/components/schemas/MemoryKind"
            },
            "type": "array",
            "title": "Kinds"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "entity_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Entity Keys"
          },
          "query_embedding": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Query Embedding"
          },
          "include_relations": {
            "type": "boolean",
            "title": "Include Relations",
            "default": true
          },
          "include_deleted": {
            "type": "boolean",
            "title": "Include Deleted",
            "default": false
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "query",
          "scope"
        ],
        "title": "SearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "SearchResponse"
      },
      "SearchResult": {
        "properties": {
          "memory": {
            "$ref": "#/components/schemas/MemoryRecord"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "related_memories": {
            "items": {
              "$ref": "#/components/schemas/RelatedMemory"
            },
            "type": "array",
            "title": "Related Memories"
          }
        },
        "type": "object",
        "required": [
          "memory",
          "score"
        ],
        "title": "SearchResult"
      },
...
      "SourceReference": {
        "properties": {
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uri"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "span_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Span Start"
          },
          "span_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Span End"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "source_type",
          "source_id"
        ],
        "title": "SourceReference"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}